Move Platform types to common utils folder.
