+++ configureWorkspace/configure.ts
@@ -15,7 +15,8 @@ import { IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
 import { globAsync } from '../helpers/async';
 import { extractRegExGroups } from '../helpers/extractRegExGroups';
-import { OS, Platform, promptForPort, quickPickOS, quickPickPlatform } from './config-utils';
+import { OS, Platform } from '../utils/platform';
+import { promptForPort, quickPickOS, quickPickPlatform } from './config-utils';
 import { configureAspDotNetCore, configureDotNetCoreConsole } from './configure_dotnetcore';
 import { configureGo } from './configure_go';
 import { configureJava } from './configure_java';
