+++ configureWorkspace/configure_dotnetcore.ts
@@ -9,7 +9,7 @@ import * as path from 'path';
 import * as semver from 'semver';
 import { extractRegExGroups } from '../helpers/extractRegExGroups';
 import { isWindows, isWindows10RS3OrNewer, isWindows10RS4OrNewer } from '../helpers/windowsVersion';
-import { OS, Platform } from './config-utils';
+import { OS, Platform } from '../utils/platform';
 import { PackageInfo } from './configure';
 
 // This file handles both ASP.NET core and .NET Core Console
