+++ test/configure.test.ts
@@ -8,8 +8,8 @@ import * as assertEx from './assertEx';
 import * as vscode from 'vscode';
 import * as fse from 'fs-extra';
 import * as path from 'path';
-import { Platform, OS } from "../configureWorkspace/config-utils";
 import { ext } from '../extensionVariables';
+import { OS, Platform } from '../utils/platform';
 import { Suite } from 'mocha';
 import { configure, ConfigureTelemetryProperties, configureApi, ConfigureApiOptions } from '../configureWorkspace/configure';
 import { TestUserInput, IActionContext, TelemetryProperties } from 'vscode-azureextensionui';
