+++ utils/platform.ts
@@ -1,0 +1,15 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE.md in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+export type OS = 'Windows' | 'Linux';
+export type Platform =
+    'Go' |
+    'Java' |
+    '.NET Core Console' |
+    'ASP.NET Core' |
+    'Node.js' |
+    'Python' |
+    'Ruby' |
+    'Other';
